Optimize fetch_clinvar_variants by batching ClinVar API calls
    • Replace per‐ID ThreadPoolExecutor esummary calls with batched requests (default batch_size=100)
    • Fetch up to 100 variant IDs at once via a single esummary call to reduce HTTP round-trips
    • Remove unnecessary threading overhead and lower overall latency when loading large variant lists
    • Preserve “Enter custom variant” option and existing caching behavior
