+++ .gitignore
@@ -1,0 +1,4 @@
+.python-version
+pyproject.toml
+README.md
+uv.lock
